Move activate/inactivate buttons to their own component
